Reorganise packages/lib
Move the majority of the script files into a src directory to be
consistent with the organisation of the other packages.
