+++ packages/lib/browser.js
@@ -1,15 +1,15 @@
-export * from "./config";
-export * from "./data";
-export * from "./errors";
-export * from "./factorio/exchange_string";
-export * from "./helpers";
-export * from "./ini";
-export * from "./link";
-export * from "./logging";
-export * from "./plugin";
-export * from "./schema";
-export * from "./users";
+export * from "./src/config";
+export * from "./src/data";
+export * from "./src/errors";
+export * from "./src/factorio/exchange_string";
+export * from "./src/helpers";
+export * from "./src/ini";
+export * from "./src/link";
+export * from "./src/logging";
+export * from "./src/plugin";
+export * from "./src/schema";
+export * from "./src/users";
 
-export { default as ExponentialBackoff } from "./ExponentialBackoff";
-export { default as PlayerStats } from "./PlayerStats";
-export { default as RateLimiter } from "./RateLimiter";
+export { default as ExponentialBackoff } from "./src/ExponentialBackoff";
+export { default as PlayerStats } from "./src/PlayerStats";
+export { default as RateLimiter } from "./src/RateLimiter";
