+++ packages/lib/index.js
@@ -4,32 +4,32 @@
  */
 "use strict";
 const libBuildMod = require("./build_mod");
-const libCommand = require("./command");
-const libConfig = require("./config");
-const libData = require("./data");
-const libDatabase = require("./database");
-const libErrors = require("./errors");
-const libFactorio = require("./factorio");
-const libFileOps = require("./file_ops");
-const libHash = require("./hash");
-const libHelpers = require("./helpers");
-const libIni = require("./ini");
-const libLink = require("./link");
-const libLogging = require("./logging");
-const libLoggingUtils = require("./logging_utils");
-const libLuaTools = require("./lua_tools");
-const libPlugin = require("./plugin");
-const libPluginLoader = require("./plugin_loader");
-const libPrometheus = require("./prometheus");
-const libSchema = require("./schema");
-const libSharedCommands = require("./shared_commands");
-const libStream = require("./stream");
-const libUsers = require("./users");
-const libZipOps = require("./zip_ops");
+const libCommand = require("./src/command");
+const libConfig = require("./src/config");
+const libData = require("./src/data");
+const libDatabase = require("./src/database");
+const libErrors = require("./src/errors");
+const libFactorio = require("./src/factorio");
+const libFileOps = require("./src/file_ops");
+const libHash = require("./src/hash");
+const libHelpers = require("./src/helpers");
+const libIni = require("./src/ini");
+const libLink = require("./src/link");
+const libLogging = require("./src/logging");
+const libLoggingUtils = require("./src/logging_utils");
+const libLuaTools = require("./src/lua_tools");
+const libPlugin = require("./src/plugin");
+const libPluginLoader = require("./src/plugin_loader");
+const libPrometheus = require("./src/prometheus");
+const libSchema = require("./src/schema");
+const libSharedCommands = require("./src/shared_commands");
+const libStream = require("./src/stream");
+const libUsers = require("./src/users");
+const libZipOps = require("./src/zip_ops");
 
-const ExponentialBackoff = require("./ExponentialBackoff");
-const PlayerStats = require("./PlayerStats");
-const RateLimiter = require("./RateLimiter");
+const ExponentialBackoff = require("./src/ExponentialBackoff");
+const PlayerStats = require("./src/PlayerStats");
+const RateLimiter = require("./src/RateLimiter");
 
 
 module.exports = {
