+++ packages/lib/src/command.js
