+++ packages/lib/src/config/classes.js
