+++ packages/lib/src/config/index.js
