+++ packages/lib/src/data/ModInfo.js
