+++ packages/lib/src/data/ModPack.js
