+++ packages/lib/src/data/messages_core.js
