+++ packages/lib/src/data/version.js
