+++ packages/lib/src/errors.js
