+++ packages/lib/src/factorio/exchange_string.js
