+++ packages/lib/src/factorio/index.js
