+++ packages/lib/src/file_ops.js
