+++ packages/lib/src/hash.js
