+++ packages/lib/src/is_deep_strict_equal.js
