+++ packages/lib/src/plugin.js
