+++ packages/lib/src/plugin_loader.js
