+++ packages/lib/src/stream.js
